+++ lib/fisk8viewer/parser/competition_summary_parser.rb
@@ -10,7 +10,7 @@ module Fisk8Viewer
         dates = []
         ary_datestr.each do |datestr|
           datestr = trim(datestr)
-          next unless datestr =~ %r{^[0-9/.]+$}
+          next unless datestr =~ %r{^[0-9/.\-]+$}
           begin
             Time.zone ||= "UTC"
             dates << Time.zone.parse(datestr)
@@ -18,6 +18,7 @@ module Fisk8Viewer
             return true
           end
         end
+        binding.pry
         return dates.max - dates.min > 3600 * 24 * 10
       end
       def parse_datetime(str, mdy_format: false)