+++ spec/controllers/skaters_controller_spec.rb
@@ -10,10 +10,13 @@ RSpec.describe SkatersController, type: :controller do
       c.start_date = "2017-1-1"
       c.end_date = "2017-1-3"
     end
-    competition.scores.create(skater: skater, category: "MEN", segment: "SHORT") do |score|
+    score = competition.scores.create(skater: skater, category: "MEN", segment: "SHORT") do |score|
       score.ranking = 1
       score.tss = 100
     end
+    score.elements.create(number: 1, name: "3T", goe: 3, base_value: 10, value: 13)
+    score.components.create(number: 1, name: "Skating Skills", value: 9)
+
     competition.scores.create(skater: skater, category: "MEN", segment: "FREE") do |score|
       score.ranking = 2
       score.tss = 200
@@ -22,6 +25,7 @@ RSpec.describe SkatersController, type: :controller do
       cr.ranking = 1
       cr.points = 300
     end
+    
   end
 
   describe 'index' do